# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/feature-flags": {
            "get": {
                "description": "Returns a list of all feature flags for the tenant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feature Flags"
                ],
                "summary": "List all feature flags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/featureflags.FeatureFlag"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new feature flag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feature Flags"
                ],
                "summary": "Create feature flag",
                "parameters": [
                    {
                        "description": "Feature flag data",
                        "name": "flag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/featureflags.FeatureFlagInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/featureflags.FeatureFlag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/feature-flags/check/{name}": {
            "get": {
                "description": "Checks if a feature flag is enabled",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feature Flags"
                ],
                "summary": "Check feature flag status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Feature flag name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/feature-flags/{id}": {
            "put": {
                "description": "Updates an existing feature flag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feature Flags"
                ],
                "summary": "Update feature flag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feature flag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Feature flag data",
                        "name": "flag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/featureflags.FeatureFlagInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/featureflags.FeatureFlag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a feature flag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feature Flags"
                ],
                "summary": "Delete feature flag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feature flag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/feature-flags/{id}/toggle": {
            "post": {
                "description": "Toggles a feature flag enabled/disabled state",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feature Flags"
                ],
                "summary": "Toggle feature flag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feature flag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/featureflags.FeatureFlag"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/feature-flags/{name}": {
            "get": {
                "description": "Returns a specific feature flag by name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feature Flags"
                ],
                "summary": "Get feature flag by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Feature flag name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/featureflags.FeatureFlag"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "featureflags.FeatureFlag": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "metadata": {
                    "description": "JSON adicional para config",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "featureflags.FeatureFlagInput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "metadata": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
...
    }
}